project_agent_stack/agents/requirements.txt
autogen
pyautogen
requests
redis
weaviate-client
python-dotenv
